refactor: collapse auth/ barrel — merge auth.ts into index.ts + self-reexport
The auth/ directory had a two-line index.ts barrel:

    export * as Auth from "./auth"
    export { OAUTH_DUMMY_KEY } from "./auth"

Since the Auth namespace and OAUTH_DUMMY_KEY both live in the same
file (auth/auth.ts), collapse the barrel into auth/index.ts directly:

  - Delete the two-line barrel
  - Rename auth/auth.ts -> auth/index.ts
  - Append `export * as Auth from "."` at the bottom

Consumers that write `import { Auth } from "@/auth"` or
`import { OAUTH_DUMMY_KEY } from "../auth"` continue to work unchanged:
both resolve to auth/index.ts which now exports both Auth (via
self-reexport) and OAUTH_DUMMY_KEY (as a top-level const).

The collapse-barrel script doesn't handle this case (it expects a
single-line `export * as X from "./y"` barrel), so this was done
manually. The spec's section on when to collapse to index.ts confirms
this is the right shape for a single-namespace directory.

Verified: tsgo --noEmit 0 errors, SDK generate no drift,
test/auth/ + test/config/config.test.ts 77/77 pass.
